Add files via upload
main py file currently obtains the listing information on the front page where it ends soonest. 
It can currently return the the time left of an item and the price of said item.
Looking to parse out the item name and couple together both the item and the price & time.